The class TrainingContext moved in separate *.py file, lr monitoring has been maintained, some log invoking have been reordered
